pyproject.toml
[tool.poetry]
name = "sps-bridge"
version = "0.0.1"
description = ""
authors = [
    "Moritz Sauter <sauter@synyx.de>",
    "David Bauer <bauer@synyx.de>",
]
license = "MIT"
readme = "README.md"


packages = [
    { include = "bridge", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = { version = "1.6.1" }
python-snap7 = "^1.1"
overrides = "^6.1.0"
Flask = "^2.1.2"
dataclasses-json = "^0.5.7"
types-requests = "^2.27.30"
StrEnum = "^0.4.7"
Flask-Cors = "^3.0.10"


[tool.poetry.dev-dependencies]
docker-compose = "1.29.2"
black = "22.3.0"
mypy = "0.961"
pylint = "2.13.8"
pyupgrade = "2.32.1"
isort = "5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry_core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint.master]
extension-pkg-whitelist = ["numpy", "cv2"]
jobs = 0

[tool.pylint.basic]
good-names = [
    "i", # iterator
    "_", "__", "___", # unneeded variable
    "T", # generic type
    "id",
]

[tool.pylint.typecheck]
# pure c module
generated-members = ["prctl.set_name"]
# necessary, at least for systems without cuda
# otherwise extension-pkg-whitelist should do the trick
ignored-modules = [
    "cv2.fisheye"
]

[tool.pylint.message_control]
disable = [
    "missing-module-docstring", "missing-class-docstring", "missing-function-docstring",
    "line-too-long",
    "too-few-public-methods",
    "duplicate-code",
    "too-many-ancestors",
]
enable = [
    "useless-suppression"
]

[tool.pylint.design]
max-attributes = 10

[tool.mypy]
python_version = "3.10"

junit_xml = "xunit-reports/xunit-result-mypy.xml"

# output style configuration
show_column_numbers = true
show_error_codes = true
pretty = true

# additional warnings
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true

no_implicit_optional = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
# disallow_any_explicit = true
disallow_any_generics = true
disallow_untyped_calls = true

# ignores that library has no typing information with it
[[tool.mypy.overrides]]
module = [
    "prctl",
    "paho.*",
    "snap7.*",
    "strenum"
]
ignore_missing_imports = true
